Invoke all `camera#` channel methods from iOS MGLMapViewDelegate
The iOS `MapboxMapController` does not invoke the channel methods
`camera#onMoveStarted`, `camera#onMove`, or `camera#onIdle`, and so
`MapboxMapController` can not notify its listeners when the map
position is moved interactively on iOS.

This changeset implements the appropriate delegate functions of the
`MGLMapViewDelegate` protocol to invoke these channel methods with the
expected arguments.

Note: the Android `MapboxMapController` provides an `isGesture` argument
to `onMoveStarted` and a `map` argument to `onIdle`, but neither of the
arguments are read when the method is handled in the Flutter
implementation of `MapboxMapController`. This iOS implementation only
provides the argument that is read on the Flutter side: the `position`
argument to `onMove`.