Fix MapboxMapsPlugin Nullpointer (#191)
* Fix MapboxMapsPlugin Nullpointer

When the app is initialized in the background the
activity provided by the Regisrar can be null,
causing the app to crash. This commit checks if
the activity is null before resolving it's hashCode.
If it is null the initialization of
this.registrarActivityHashCode is delayed until
onActivityCreated is called.

* Revert "Fix MapboxMapsPlugin Nullpointer"

This reverts commit 2e4371400a3ce29c627048328350fec20c40336a.

* Only Register in Foreground

When registering the app check if the activity != null.
If it is null cancel the initialization as the plugin is foreground only.

Fix is based on the GoogleMapsPlugin.