Provide `onMapIdle` callback
Corresponding to `mapViewDidBecomeIdle` in the iOS SDK, this informs the
application "that the map view is entering an idle state, and no more
drawing will be necessary until new data is loaded or there is some
interaction with the map."

https://docs.mapbox.com/ios/api/maps/5.6.1/Protocols/MGLMapViewDelegate.html#/c:objc(pl)MGLMapViewDelegate(im)mapViewDidBecomeIdle:

I have not provided an Android implementation as I am not set up to test
one yet. I would do this later if it's still missing.