Add First-Class Support For Listening to `onCameraIdle` events
I believe the use case for the `onCameraIdle` event [1] is large enough
that this event should be able to be listened to directly without having
to listen to extraneous events via the ChangeNotifier interface. The
extraneous events which I think add too much noise when combined with
the camera idle event include:

* Camera Start/Move Events
* Updating Map Options
* Adding/Updating/Removing/Clearing a Symbol, Line, or Circle

For example, the use case described in #131 where a client wants to
display symbols within a bounding box after a user pans the map should
be able to have first-class support. FWIW, other flutter mapping plugins
provide this ability, and it is very useful [2].

Usage:

```
MapboxMap(
  onCameraIdle: () {
    print("Camera has idled.");
  },
);
```

[1] https://github.com/tobrun/flutter-mapbox-gl/blob/1bbe7ea7adc8030af16a3cdbf5ade2f94ddea6bd/lib/src/controller.dart#L156
[2] https://github.com/flutter/plugins/blob/61775d32f1d1f0997e0e93bb1e8676b96be9d848/packages/google_maps_flutter/google_maps_flutter/lib/src/google_map.dart#L136