Avoid white space when resizing map on web
When resizing the browser screen Flutter fires several resize events.
These events are properly propagated to the containing div, but not
always by the map. It appears Mapbox misses the last resize event.
Therefore, we fire a timer that will check for mismatching width/height
between mapbox and it's container and trigger an additonal resizing.