Update tracked camera position in camera#onIdle (#500)
When animating to a new camera position, the final resting position of
the camera is generally not received in a `camera#onMove` callback. The
latest camera position known to the flutter-mapbox-gl framework and the
underlying mapbox-gl framework is therefore out of sync, until the next
time the camera is moved. This same issue would exist when the camera is
moved with a swipe, but it's less prominent since the camera is not
moving as quickly.

The `camera#onIdle` callback is a good place to update the camera
position, since it is at rest and no longer moving. This insures that
the framework has an accurate camera position if the client application
requests one from its own `onCameraIdle` callback, or atany other time
between camera moves.