Adapt example to OfflineRegionDefinition class
This was fairly tricky as the example depends on a fixed set of regions
where the IDs were decided at compile time and expected to remain the
same after being downloaded, which is no longer the case for Android or
iOS. (This example would compile before the dart `OfflineRegion`
refactoring, but it wouldn't have worked.)

The most straightforward way to adapt it was to add a metadata field
"name" which serves to identify elements of this fixed set of regions
before and after they're saved. This feels contrived, but the case
demonstrated in the example where the downloadable regions are fixed in
advance is itself rather unusual.

Also worth noting that this example does not wait for the download to
complete before indicating that it was downloaded, it just waits for the
channel response from the download request message. I've rearranged the
code to read in the generated id from that channel response, but the
behavior is the same as before.