Support override of attribution click action (iOS)
This override allows the application to provide a custom UX for Mapbox
telemetry settings and all required copyright information.

This improves the utility of the existing `getTelemetryEnabled` and
`setTelemetryEnabled` methods, as there is otherwise no means to provide
an alternative to the basic action sheet shown by default.

While it seemed a bit dubious to alter the click target of a UI element
that belongs to the underlying framework, it's apparently expected that
the application may override the button's visibility as that use case is
handled by the framework. (The application is prevented from running
unless a particular property is added to the info.plist.)

I explored that route as well, hiding the attribution button and
replacing it with an imitation, however the built-in button has features
that would be difficult to replicate in an externally added view. For
example, it repositions itself when the content insets change, and it's
always properly aligned with the Mapbox logo to the left.

Since the appearance and positioning of the built-in button is good and
appropriate for almost any application, but the action sheet is rather
basic and won't include whatever additional attributions an application
may require, overriding its click action appears the be the best option.