Add attribution button gravity, position normally
This continues the work of PR #615 to fix #261 and #540, and replaces
the changed calculation of #681 with a more general solution.

The underlying issue here is that the attribution button is left-aligned
by default on Android, and right-aligned by default on iOS. The approach
taken in #681 where the margin x value is assumed to represent the left
margin does not generalize to iOS, where it will be a right margin.

Since the alignment / gravity of the button can be configured on each
platform, we can address this problem at its source by exposing this
configuration through flutter-mapbox-gl. Then on Android, we can apply
the button margins same way the compass is positioned, with respect to
its gravity.

I've tested this on iOS and Android with both bottom left and bottom
right attribution buttons. The best option currently is to configure the
attribution button to bottom right, where it falls by default on iOS. In
this way you will have a similar experience on both platforms with a
fully configurable margin.

The mapbox logo's alignment/position is also configurable, should anyone
want to expose it in future.