Improve switch default behavior
When setting the position of the compass and attribution button, we need
to assume the platform default gravity for those elements if no explicit
value has been provided. An incorrect default was assumed for the
attribution button (copied from the compass), so that is corrected here.

The methods for setting gravity do not require a default since they are
only called if an explicit gravity value has been provided. Those
default cases are removed here.

https://github.com/tobrun/flutter-mapbox-gl/pull/731#discussion_r740930948