Add attribution button gravity, position normally  (#731)
* fix(android): allow android logoViewMargins and attributionButtonMargins to behave similarly to ios

* fix(convert): compass view margin remain the same since theres the gravity and it cant be tested now

* feat(setAttributionButtonMargins): adding variable name to the constant and adjust it to density

* Add attribution button gravity, position normally

This continues the work of PR #615 to fix #261 and #540, and replaces
the changed calculation of #681 with a more general solution.

The underlying issue here is that the attribution button is left-aligned
by default on Android, and right-aligned by default on iOS. The approach
taken in #681 where the margin x value is assumed to represent the left
margin does not generalize to iOS, where it will be a right margin.

Since the alignment / gravity of the button can be configured on each
platform, we can address this problem at its source by exposing this
configuration through flutter-mapbox-gl. Then on Android, we can apply
the button margins same way the compass is positioned, with respect to
its gravity.

I've tested this on iOS and Android with both bottom left and bottom
right attribution buttons. The best option currently is to configure the
attribution button to bottom right, where it falls by default on iOS. In
this way you will have a similar experience on both platforms with a
fully configurable margin.

The mapbox logo's alignment/position is also configurable, should anyone
want to expose it in future.

* Improve switch default behavior

When setting the position of the compass and attribution button, we need
to assume the platform default gravity for those elements if no explicit
value has been provided. An incorrect default was assumed for the
attribution button (copied from the compass), so that is corrected here.

The methods for setting gravity do not require a default since they are
only called if an explicit gravity value has been provided. Those
default cases are removed here.

https://github.com/tobrun/flutter-mapbox-gl/pull/731#discussion_r740930948

* Rename "gravity" methods in web implementation

This name is not commonly used outside of android. Also, we're
converting to the typed position objects as quickly as possible with
these changes, rather than passing around the int.

https://github.com/tobrun/flutter-mapbox-gl/pull/731#discussion_r741027579

Co-authored-by: Ching Yaw Hao <ching.yh@mydidian.com>